requirements.txt
certifi==2018.4.16
chardet==3.0.4
idna==2.7
influxdb==5.2.0
paho-mqtt==1.3.1
pkg-resources==0.0.0
pyserial==3.4
python-crontab==2.3.4
python-dateutil==2.7.3
pytz==2018.5
PyYAML==3.13
requests==2.19.1
setuptools-scm==3.0.4
six==1.11.0
urllib3==1.23
